extract Entitlement class to a separate file
